# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para operaciones del sistema SistaL.",
        "title": "MS Operations API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8083",
    "basePath": "/api/v1",
    "paths": {
        "/despachos": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene lista paginada de despachos con filtros opcionales",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Despachos"
                ],
                "summary": "Obtener despachos",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Número de página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Elementos por página",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filtrar por ID de estado",
                        "name": "id_estado_despacho",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por sucursal",
                        "name": "sucursal",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha desde (YYYY-MM-DD)",
                        "name": "fecha_desde",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha hasta (YYYY-MM-DD)",
                        "name": "fecha_hasta",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Búsqueda en guía o voucher",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de despachos con paginación",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea un nuevo despacho en el sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Despachos"
                ],
                "summary": "Crear despacho",
                "parameters": [
                    {
                        "description": "Datos del despacho",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateDespachoDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Despacho creado exitosamente",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Datos de entrada inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/despachos/{id_despacho}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene detalles completos de un despacho específico",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Despachos"
                ],
                "summary": "Obtener despacho por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del despacho",
                        "name": "id_despacho",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del despacho",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Despacho no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Actualiza campos de un despacho existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Despachos"
                ],
                "summary": "Actualizar despacho",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del despacho",
                        "name": "id_despacho",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos a actualizar",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateDespachoDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Despacho actualizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Datos de entrada inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Despacho no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/despachos/{id_despacho}/estado": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cambia el estado de un despacho",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Despachos"
                ],
                "summary": "Cambiar estado de despacho",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del despacho",
                        "name": "id_despacho",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nuevo estado",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CambiarEstadoDespachoDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado actualizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Transición de estado inválida",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Despacho no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.CambiarEstadoDespachoDTO": {
            "type": "object",
            "required": [
                "id_estado_despacho"
            ],
            "properties": {
                "id_estado_despacho": {
                    "type": "integer"
                },
                "responsable_recepcion": {
                    "type": "string"
                }
            }
        },
...
        "handlers.CreateDespachoDTO": {
            "type": "object",
            "required": [
                "guia_de_despacho",
                "sucursal"
            ],
            "properties": {
                "fecha_despacho": {
                    "type": "string"
                },
                "guia_de_despacho": {
                    "type": "string"
                },
                "id_estado_despacho": {
                    "type": "integer"
                },
                "id_factura": {
                    "type": "integer"
                },
                "numero_voucher": {
                    "type": "string"
                },
                "responsable_recepcion": {
                    "type": "string"
                },
                "sucursal": {
                    "type": "string"
                }
            }
        },
...
        "handlers.UpdateDespachoDTO": {
            "type": "object",
            "properties": {
                "id_estado_despacho": {
                    "type": "integer"
                },
                "id_factura": {
                    "type": "integer"
                },
                "numero_voucher": {
                    "type": "string"
                },
                "responsable_recepcion": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Token de autenticación JWT en formato Bearer",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}